requirements.txt
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
h11==0.14.0
idna==3.10
oauthlib==3.2.2
outcome==1.3.0.post0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.38.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.0
websocket-client==1.8.0
wsproto==1.2.0
